website/requirements.txt
plotly
pandas
pyyaml